changes for Voellmy-Salm and pore pressure
Changes in collaboration with Eric Breard.
The effect of pore pressure of friction term has been rewritten, now with an additional check on the positivity of the term corresponding to Eq. 2 from Gueugneau et al. 2017
The effective friction coefficient mu_eff is computed in inpout_2d.f90 and saved in p_2d files.
In the example xi has been increased to make it more effective the effect of pore pressure on the dynamics.